+++ examples/lottie/src/main.rs
@@ -19,11 +19,10 @@ fn main() {
 fn load_lottie(mut commands: Commands, asset_server: ResMut<AssetServer>) {
     commands.spawn((Camera2d, VelloView));
 
-    // Yes, it's this simple.
-    commands.spawn(VelloLottieBundle {
-        asset: VelloLottieHandle(asset_server.load("embedded://lottie/assets/Tiger.json")),
-        debug_visualizations: DebugVisualizations::Visible,
-        transform: Transform::from_scale(Vec3::splat(0.5)),
-        ..default()
-    });
+    // You can also use `VelloLottieBundle`
+    commands
+        .spawn(VelloLottieHandle(
+            asset_server.load("embedded://lottie/assets/Tiger.json"),
+        ))
+        .insert(Transform::from_scale(Vec3::splat(0.5)));
 }