+++ examples/scene_ui/src/main.rs
@@ -30,7 +30,7 @@ fn setup_ui(mut commands: Commands) {
         },
         BorderColor(css::FUCHSIA.with_alpha(0.5).into()),
         Interaction::default(),
-        VelloScene::default(),
+        VelloScene::new(),
         CoordinateSpace::ScreenSpace,
     ));
 }