+++ src/text/vello_text.rs
@@ -1,8 +1,15 @@
-use crate::VelloFont;
+use crate::{debug::DebugVisualizations, CoordinateSpace, VelloFont};
 use bevy::prelude::*;
 use vello::peniko::{self, Brush};
 
 #[derive(Component, Default, Clone)]
+#[require(
+    VelloTextAnchor,
+    CoordinateSpace,
+    Transform,
+    DebugVisualizations,
+    Visibility
+)]
 pub struct VelloTextSection {
     pub value: String,
     pub style: VelloTextStyle,